# mvcc: remove allocation from Put path in tree index