Before
```
BenchmarkStorePut-4                               200000              6853 ns/op
BenchmarkStorePutUpdate-4                         200000              6102 ns/op
BenchmarkStoreTxnPut-4                            200000              6794 ns/op            1490 B/op         15 allocs/op
BenchmarkWatchableStorePut-4                      200000              6743 ns/op            1550 B/op         15 allocs/op
BenchmarkWatchableStoreTxnPut-4                   200000              7341 ns/op            1563 B/op         17 allocs/op
BenchmarkWatchableStoreWatchSyncPut-4             200000              6057 ns/op             275 B/op          2 allocs/op
PASS
ok      _/Users/anthony/go/src/github.com/coreos/etcd/mvcc      17.834s
```

After
```
BenchmarkStorePut-4                               300000              3461 ns/op
BenchmarkStorePutUpdate-4                         200000              6256 ns/op
BenchmarkStoreTxnPut-4                            300000              3353 ns/op            1393 B/op         14 allocs/op
BenchmarkWatchableStorePut-4                      500000              3315 ns/op            1436 B/op         14 allocs/op
BenchmarkWatchableStoreTxnPut-4                   300000              3631 ns/op            1454 B/op         16 allocs/op
BenchmarkWatchableStoreWatchSyncPut-4             200000              5916 ns/op             271 B/op          2 allocs/op
PASS
ok      _/Users/anthony/go/src/github.com/coreos/etcd/mvcc      26.034s
```