@heyitsanthony 2017-06-06
turns out this is broken; I'm surprised it's passing the cov checks since `TestIndexGet` fails every run. Travis also seems to be missing tests. Likely due to shellcheck churn.